move container logic into its own file
